requirements.txt
requests_toolbelt==0.9.1
fastapi==0.95.1
uvicorn[standard]==0.22.0
firebase-admin==6.0.1
pyrebase4==4.5.0
boto3==1.26.158
pydantic==1.10.7
celery==5.3.0
redis==4.5.1
opencv-python-headless==4.5.5.64
tensorflow==2.16.1
keras==3.0.0
pillow==9.5.0
dill==0.3.4
python-multipart==0.0.6